Refactor to perform launch preparation in common method
